Completed Search
Completed search to a state that is almost done, maybe more features in the future and documentation, but for now is done.
Can search based of the name of the file.
Can search based of the content in the file.
Can search based on an entire directory.
Can search in inverse which allows searching of nots rether than is.
Can search with deep searches which includes all folders inside a parent folder.
Can search in contain groups to allowing searching of fields that all share the contains.
Fixed bug with search directory.