Updated Textures and Results
Reformatted all of results and colors, to be organized differently and all be inside the display import, this including calling created results.

Renamed the results from file and text to outer and inner.

Polished up the colors for results and errors.

Allow textures to now change multiple files at once, commonly will be used with the search methods.

Added new texture altering methods, including: invert, blur, resize, and noise.
Removed tint for now, might bring back.

Noise might need to be altered, is working, but now how i envisioned or want it to.