Small changes
Forgot to add chances into images and plan to add them to all methods, i do understand that they will not be the most used, but for how simple they are to add. they are worth it if someone wants to goof around with them.

Started work on outer first, currently working on normal.

improved the .gitingore by removing files that were in the past pushed into the repository.