Outer
Worked on most of Outer normal method, as everything but the extra parameters possible with randomization should be working, however have not have time to really test it out.

also removed duplicated, flatten, and weight from the parameters as i believe one or two people could find it useful, but its not worth the amount of other problems it creates, such as complexity and possible speed.

Additionally changed uuid4 to hashes as they are majorly faster, however they are more likely to collide at bigger groups, such as 5-10 million or higher, but this should not be a problem for 99 percent of people, might make a safety mode, but i am not sure yet.