Outer work
Finished normal and group shuffling.
Removed flags form normal and added contains in group
Overall improvements from older version are that you can now use presets and preshuffles, as well as filters
Fixed display results length in both otter and images, and also future methods

With the addition of reshuffles, there is no need for reverse or rotate. as you can do this with preshuffles, i will just need to make sure this is documented correctly to explain to others.