Audio and inner
Setting up general lines of code for inner and outer methods, however have not stated to real code into the methods.

Added the Audio file to manage audio, i am still going to use librosa as even if it is way less powerful than other options, but those require alternative downloads.

Worked on 3  methods for audio, which are: volume, pitch, and tempo.

Removed the rotation and reverse methods from inner as they are not useful with preshuffle.

Added chance_contains within inner methods