New problems to solve
In __init__ made it so that it was more obvious and bold what random was used for.

Worked on Inner and mode some progress but some problems occured.

As I worked on Inner and dealt with preshufle, presets, and chances. I relized how complicared it will be to work all together and I also plan to step back and see how to deal with this, as i see both as really important.

Also plan to work again on Outer, as i see a problem could occur with filters, I found this when working on Inner