Planning and changes
Giving some time to thing, for now, it is not worth the difficult that comes with presets as it dose not have a massive use that can not be done with a little more complex with other methods. additionally it brings difficulties to other features I plan to add. Additionally, inner and outer would use it differnetly and even different methods inside outer, so it could be confusing. With all of that in mind it will be shelfed and look at possibly at a later data.

Fixed a small bug with method methods and sending errors in string formats instead of error formats.

Restricted groups in outer to make sure it works correct with multiple and more complex filters. I fixed this by changing how loops are used with files and contains.

Fixed some small comments due to saying presets when it should of said preshuffle.

fixed a typo where the output of groups said normal instead of groups.