INNER
working on normal inner and beleive i have found a better way to do all of this, however to make sure i do not mess up everything i am pushing this to save it all. the only think left to deal with is chance file, and i think the new way with classes will be better. one i test this and push which ever way works, I will update and log all the changes to normal inner