Inner normal
Inner normal is at a state that i like it for now, it is up to pare and better than the original.

Compared to the original, i added filters, flatening, duplication, presets, and preshuffles. Additionally i added 4 total chance types total chance, data chance, contain chance, and file chance.

However weight is not at a good state, as it requires users to know exactly how many options will be shuffled, I am still keeping it in, however its not at the best state and will be worked on later.

Also found a way to make presets to be consistant with each method, at least for Inner methods, not sure about Outer yet.

Also slightly worked on the methods parameters that will be used in scale and offset. Additionally, i will most likely work on those as they are the easier that group Inner, let god help me on Inner group.