Scaling and Offsets
Scaling and offset method have been added and massively improved from there original version.

We added many new parameters to allow for better control. this includes outter and inner clamps which allow you to control the range of data can be created. zeros which will make sure the output cant be zero, rounding to round the decimals, and the ability to turn on and off decimals. Then there is minmax which will allow for groups of numbers that require one to be bigger than the other.

range is how you determine which numbers can be selected with options for int or float. However, how i do it now can be a little confusion, so either i need to change how i do it or document it well.

also added rerolling of chances to make sure the above criteria's don't make data impossible to be created, however if it fails, it will just not change the output and warn the users

I can tell that small things can still be improved with them, such as when warnings and outputs are displayed, and I still need to make it so that weight will work for normal.

However for now I plan to add 2 new methods, with the goal of displaying what data it would find in group and normal, allowing users to better calculate presets, preshuffles, and weight, however this will be done after group inner.
This allows me to not be restrictive with how weight and others are used as i had to worry about user putting the correct amount, as now i can Trust them to look into the data.

That was a lot but now i have to work on the worse one, Group Inner, this will be not fun.