Procrastination
Not gonna lie, progracinating the big things i have to do, which is finishing group inner, but i have done many important things.

I have done some work on group inner, but need to finish it up

Worked on a layout for the final test file for the shuffler, and listed changes i plan to go though and need to finish in that while testing.

I am having to farm and work on stuff for research, so even though i am more active that i have been in a while, it will still be kinda slow for now. For now I plan to chip at the list first or finish group, which ever i feel like is best for now.

Also working on polishing stats as there was a bug with all results sharing stats and made a better way to display a classes stats with result_stats.