Searches, Outer, and Inner
Got many things worked on and progress on many other things:
Draft template for method docstrings currently only inside Outer, still working on format.
Also remoted flags from outer group, as based on what i remember and looked into, it really is useless and could cause problems, making it useful in any kind of way would require me to reformat how the parameters are, and i would rather keep consistency over something that would add essentially nothing new.

Still working on Inner group, however trying to keep consistency and all of that is causing complexation, however still working on it.
Also removed the outputs from inner and moved them to search.

Search has 2 new methods and might get more if they are needed for future random categories in the future. Inner and Outer search are used to get the data that would be shuffled in normal group and normal methods, but does not alter files, it allows users to have better times with managing and working with weight and preshuffle, as they are working less blind. To activate the grouping or non grouping versions is as simple as a Boolean parameter.

I do understand that outer search is essentially useless, however i am petter and need order, so i am doing it  so that all other shuffle types can have one. also some people might not realize it is useless, allowing for more inexperience users to use this first and realize later about better options.

Also removed and added things that need to be worked on in the bottom of Randomize.py