Small changes and big problems
Worked on small changes that needed work but new problems are occuring.

Worded on documentation format and added some small template documentation to some other methods.
Fixed a bug with source length not always updating, now it always resets to 0 after a method is done
Worked on the foundation of sheets of images, and have created the innit method to create the images form one sheet.

However i have had ALOT of problems setting up sheets to my likings, as i just feel its not coming to my liking or to the style of other methods, much more time will need to be spent on it.

Also added more things to be worked on and realized there is no clean way to do preshuffles as its just so messy to do and feel its not worth it.