Image Sheets
Got image sheets to finally, work, insted of making it so that there is a method for it, I have altered other methods to allow for a combination of methods to let it happen, I do believe this is better overall. To do this i create the create method in image class, alowing you to create new photo files, and i updated layer and crop.

Update the feature list of RANDOMIZE.

did some small work to inner group, but am still working on it.

removed zip import form methods

made the coding style of filters seem more similar to other methods