Outer Preset
Presets now work with outer group, as it will allow users to do presets for the contains being shuffled, however i have no plan for outer normal, as that would require you to enter entire file directories per file, as you are allowed to edit files between multiple directories. I could make it so you can only change the files name, so i might look into it, however it is not a priority.

Added alter filters to the inner search, to keep them similar to the original methods, and some people might relay on the new data created from the alters to be shuffled.

Also updated documentation of thigs to do in the randomize file

Currently plan to work on audio method.
This method will generally not be given as much attention as i believe it will not be used much for multiple reason, such as how long it will take each file to be altered, the size added to mods and addons, and the requirement to install ffmpeg, which users will need it to run a random generator that uses audio files. Also it seems like a complex nightmare. I am going to try, but I cannot promise it will turn out with much utility, or even come out at all.