Audio and contains
Before the main part of this code, listing some small things.

Decided that I am ok with how the library handles logs, as i will count on the user to manage the file before or after, as the libraries goal will only be to  determine if the user wants to log data, and where the data will be logged.

I have decided that i will most likely have this be version 2 of the program.
there are still many features missing such as input working, weights, or contains, however for now i am ok. This program has made alot of progress and is much more polished than the past one. For now i plan to test the library and see what future tools might be needed or what needs to be improved. So for now i leave it at this.

Still deciding how much i will document it as i am not truly happy showing this to the public yet, so who knows, for now, its time for testing.