Clean up
About to start testing so i am cleaning up the Github.

Something else to note is that I do plan to update problems with this test and then document them and what i need to work on in version 2 updates or with the brand new version. After that I will documentation as much as i can and hopefully have this version done before I start school.