Testing
Doing alot of test work with the  RANDOMIZER.py, and finding things i want to add in the future and new problems i never knew existed.

Fixed an error with display and displaying errors correctly.
Fixed a bug with filters not correctly detecting file similarities sometimes.

Changed inner scale and offset, so it uses uniform random if no mode is being used.